Revert "int compare fix."
This reverts commit 7f3711d95561db073ad6af4d0322ff26b863adbd.

+++ cpp/dolfin/mesh/Topology.cpp
@@ -117,7 +117,7 @@ void Topology::set_global_indices(
 //-----------------------------------------------------------------------------
 void Topology::set_ghost_offset(int dim, std::int32_t index)
 {
-  assert(dim < (int)_ghost_offset.size());
+  assert(dim < _ghost_offset.size());
   _ghost_offset[dim] = index;
 }
 //-----------------------------------------------------------------------------
